Add active_span method to Tracer
Previously it was only defined in OpenTracing module. Tracer#active_span
is also defined in opentracing python and java. Define it here as well
and make OpenTracing to delegate to Tracer instead.

This is useful when Tracer object is being passed around and OpenTracing
module is not used directly.

This is a backwards compatible change.